requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.7
django-debug-toolbar==3.2.2
django-debug-toolbar-template-profiler==2.0.2
django-environ==0.8.1
django-extensions==3.1.5
idna==3.3
install==1.3.5
oauthlib==3.1.1
Pillow==8.3.2
pycparser==2.21
pygraphviz==1.7
PyJWT==2.3.0
python-memcached==1.59
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.7
wrapt==1.13.3
